Anyone can create games and be an admin
Admin feature that depends on the user is now connect to the game,
thus, the person who creates the game gains the admin's abilities.
Due to this, there is no more need for ensureAdmin used on the backend.

Which game will appear on the admin tab or on the home tab
is provide with the getters on the frontend, the games that
appear on the admin are games created by the user. While we have
the games that we are admin, the socket connection
is also being established here.